requirements.txt
langchain
langchain-core
langchain-community
langchain-openai
langchain-chroma
langchain-huggingface
langchain-text-splitters
openai
numpy
chromadb
sentence-transformers
torch
python-dotenv
pypdf                      
markdown                          
pytest
pytest-mock
fastapi
uvicorn
sse-starlette
pywebview